submissions/ai_apps/ltm920716/requirements.txt
reflex>=0.2.9
openai
loguru
python-dotenv 